代码修改

+++ oceanbase-system/src/main/java/com/sinovatio/modules/system/rest/UserController.java
@@ -4,15 +4,20 @@ import com.sinovatio.aop.log.Log;
 import com.sinovatio.config.DataScope;
 import com.sinovatio.domain.Picture;
 import com.sinovatio.domain.VerificationCode;
-import com.sinovatio.modules.system.domain.User;
 import com.sinovatio.exception.BadRequestException;
+import com.sinovatio.modules.system.domain.Role;
+import com.sinovatio.modules.system.domain.User;
 import com.sinovatio.modules.system.service.DeptService;
-import com.sinovatio.service.PictureService;
-import com.sinovatio.service.VerificationCodeService;
-import com.sinovatio.utils.*;
+import com.sinovatio.modules.system.service.RoleService;
 import com.sinovatio.modules.system.service.UserService;
 import com.sinovatio.modules.system.service.dto.UserDTO;
 import com.sinovatio.modules.system.service.query.UserQueryService;
+import com.sinovatio.service.PictureService;
+import com.sinovatio.service.VerificationCodeService;
+import com.sinovatio.utils.EncryptUtils;
+import com.sinovatio.utils.OceanMobileConstant;
+import com.sinovatio.utils.PageUtil;
+import com.sinovatio.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -24,7 +29,9 @@ import org.springframework.util.ObjectUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
 * @ClassName: UserController
@@ -51,6 +58,9 @@ public class UserController {
 
     @Autowired
     private DeptService deptService;
+
+    @Autowired
+    private RoleService roleService;
 
     @Autowired
     private VerificationCodeService verificationCodeService;
@@ -84,7 +94,7 @@ public class UserController {
             if(result.size() == 0){
                 return new ResponseEntity(PageUtil.toPage(null,0),HttpStatus.OK);
             } else return new ResponseEntity(userQueryService.queryAll(userDTO,result,pageable),HttpStatus.OK);
-        // 否则取并集
+            // 否则取并集
         } else {
             result.addAll(deptSet);
             result.addAll(deptIds);
@@ -99,6 +109,7 @@ public class UserController {
         if (resources.getId() != null) {
             throw new BadRequestException("A new "+ ENTITY_NAME +" cannot already have an ID");
         }
+        checkLevel(resources);
         return new ResponseEntity(userService.create(resources),HttpStatus.CREATED);
     }
 
@@ -106,6 +117,7 @@ public class UserController {
     @PutMapping(value = "/users")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_EDIT')")
     public ResponseEntity update(@Validated(User.Update.class) @RequestBody User resources){
+        checkLevel(resources);
         userService.update(resources);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
@@ -114,38 +126,44 @@ public class UserController {
     @DeleteMapping(value = "/users/{id}")
     @PreAuthorize("hasAnyRole('ADMIN','USER_ALL','USER_DELETE')")
     public ResponseEntity delete(@PathVariable Long id){
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
+        Integer optLevel =  Collections.min(roleService.findByUsers_Id(id).stream().map(Role::getLevel).collect(Collectors.toList()));
+
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
         userService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
 
     /**
      * 验证密码
-     * @param pass
+     * @param user
      * @return
      */
-    @GetMapping(value = "/users/validPass/{pass}")
-    public ResponseEntity validPass(@PathVariable String pass){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
+    @PostMapping(value = "/users/validPass")
+    public ResponseEntity validPass(@RequestBody User user){
+        UserDetails userDetails = SecurityUtils.getUserDetails();
         Map map = new HashMap();
         map.put("status",200);
-        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
-           map.put("status",400);
+        if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
+            map.put("status",400);
         }
         return new ResponseEntity(map,HttpStatus.OK);
     }
 
     /**
      * 修改密码
-     * @param pass
+     * @param user
      * @return
      */
-    @GetMapping(value = "/users/updatePass/{pass}")
-    public ResponseEntity updatePass(@PathVariable String pass){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
-        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(pass))){
+    @PostMapping(value = "/users/updatePass")
+    public ResponseEntity updatePass(@RequestBody User user){
+        UserDetails userDetails = SecurityUtils.getUserDetails();
+        if(userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
             throw new BadRequestException("新密码不能与旧密码相同");
         }
-        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(pass));
+        userService.updatePass(userDetails.getUsername(),EncryptUtils.encryptPassword(user.getPassword()));
         return new ResponseEntity(HttpStatus.OK);
     }
 
@@ -156,9 +174,8 @@ public class UserController {
      */
     @PostMapping(value = "/users/updateAvatar")
     public ResponseEntity updateAvatar(@RequestParam MultipartFile file){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
-        Picture picture = pictureService.upload(file,userDetails.getUsername());
-        userService.updateAvatar(userDetails.getUsername(),picture.getUrl());
+        Picture picture = pictureService.upload(file, SecurityUtils.getUsername());
+        userService.updateAvatar(SecurityUtils.getUsername(),picture.getUrl());
         return new ResponseEntity(HttpStatus.OK);
     }
 
@@ -171,7 +188,7 @@ public class UserController {
     @Log("修改邮箱")
     @PostMapping(value = "/users/updateEmail/{code}")
     public ResponseEntity updateEmail(@PathVariable String code,@RequestBody User user){
-        UserDetails userDetails = SecurityContextHolder.getUserDetails();
+        UserDetails userDetails = SecurityUtils.getUserDetails();
         if(!userDetails.getPassword().equals(EncryptUtils.encryptPassword(user.getPassword()))){
             throw new BadRequestException("密码错误");
         }
@@ -179,5 +196,17 @@ public class UserController {
         verificationCodeService.validated(verificationCode);
         userService.updateEmail(userDetails.getUsername(),user.getEmail());
         return new ResponseEntity(HttpStatus.OK);
+    }
+
+    /**
+     * 如果当前用户的角色级别低于创建用户的角色级别，则抛出权限不足的错误
+     * @param resources
+     */
+    private void checkLevel(User resources) {
+        Integer currentLevel =  Collections.min(roleService.findByUsers_Id(SecurityUtils.getUserId()).stream().map(Role::getLevel).collect(Collectors.toList()));
+        Integer optLevel = roleService.findByRoles(resources.getRoles());
+        if (currentLevel > optLevel) {
+            throw new BadRequestException("角色权限不足");
+        }
     }
 }
